Include full path in error message when required field is missing
With input of

    struct {
        A struct {
            B string `required:"true"`
        }
        C struct {
            B string `required:"true"`
        }
    }

if either of `A.B` or `C.B` was unset in input, the error message
would be the same:

> B is required, but blank

and it's hard to see *which* of `A.B` or `C.B` is missing. Modify the
error message to include the full "path" to the field, so that the
message reads:

> configor.A.B is required, but blank

I haven't stripped off the prefix in this patch, so the error message
will include the result of `getPrefix`.