Fix parsing environment variables into maps
Env vars setting maps using JSON syntax would have
the value appended to the map instead of overwriting
the map.

E.g. ENV_VAR={"a":"b"} would append the "a" key to
a map[string]string map, instead of overwriting the
map to only contain the "a" key and "b" value.